backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
ultracontext
matplotlib
numpy
sympy
scipy
